requirements.txt
beautifulsoup4
requests
mysql-connector-python
python-dotenv
pandas
